Beautiful structure and Protocol
Remove unnecessary init parameter in Message in order to make a better
 extensibility.
More tests.

Right now working on seperating peer operation method and protocol ope-
ration method from Peer, sometimes varaible is only work for specific
protocol, so made them all define in Peer.py seems a little bit waste.

And yes, until this commit, the framework -- Even it is still not like a
 framework, lol -- do not have ability to execute a TCP hole punching,
and this will be our short-term goals.

DNS sync is already been prove by another service host on GCP.
But due to hole punching is still on progress, we only able to send a
handshake packet, but not a stable connection.